# Morocco-Now/lib/api-zod/src/generated/types/aiQuestion.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Morocco Now public discovery and account API
 * OpenAPI spec version: 0.1.0
 */

export interface AiQuestion {
  /** @minLength 2 */
  question: string;
  language?: string;
}