# Morocco-Now/lib/api-zod/src/generated/types/mapMarker.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Morocco Now public discovery and account API
 * OpenAPI spec version: 0.1.0
 */

export interface MapMarker {
  id: number;
  slug: string;
  name: string;
  category: string;
  city: string;
  latitude: number;
  longitude: number;
}