# Morocco-Now/lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Morocco Now public discovery and account API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  description: string;
  priceCents: number;
  currency: string;
  /** @nullable */
  imageUrl?: string | null;
  status: string;
}