# Morocco-Now/lib/api-zod/src/generated/types/review.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Morocco Now public discovery and account API
 * OpenAPI spec version: 0.1.0
 */

export interface Review {
  id: number;
  businessId: number;
  userName: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  title: string;
  comment: string;
  createdAt: string;
}