# Morocco-Now/lib/api-zod/src/generated/types/reviewInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Morocco Now public discovery and account API
 * OpenAPI spec version: 0.1.0
 */

export interface ReviewInput {
  businessId: number;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  /** @minLength 2 */
  title: string;
  /** @minLength 10 */
  comment: string;
}